Byron

+++ server/views/property_view.py
@@ -1,11 +1,14 @@
 from flask import Blueprint,jsonify,request
 from models import db, Property,User
 from sqlalchemy import desc
+from flask_jwt_extended import  get_jwt_identity, jwt_required
 
 prop_bp = Blueprint('prop_bp', __name__)
 
 #view all properties
 @prop_bp.route('/properties')
+@jwt_required()
+
 def get_all_properties():
    properties = Property.query.all()
    
@@ -40,6 +43,8 @@ def get_all_properties():
 
 #view single property
 @prop_bp.route('/properties/<int:property_id>')
+@jwt_required()
+
 def get_single_property(property_id):
    property = Property.query.filter_by(id=property_id).first()
    if not property:
@@ -73,6 +78,7 @@ def get_single_property(property_id):
 
 #search property by title
 @prop_bp.route('/properties/<string:search>', methods=['GET'])
+@jwt_required()
 def search_property(search):
    properties= Property.query.filter(Property.title.ilike(f'%{search}%')).all()
      
@@ -108,6 +114,7 @@ def search_property(search):
 
 #add property
 @prop_bp.route('/properties',methods=['POST'])
+@jwt_required()
 def add_property():
     data = request.get_json()
     
@@ -125,7 +132,8 @@ def add_property():
     beds = data['beds']
     location = data['location']
     
-    user_id=2 #current user id
+    user_id= get_jwt_identity()               #current user id
+
     user=User.query.filter_by(id=user_id).first()
     if user.user_type == 'guest':
        return jsonify({"error":"unauthorized access"}),404
@@ -184,10 +192,14 @@ def add_property():
 
 #update property
 @prop_bp.route("/properties/<int:property_id>", methods=['PATCH'])
+@jwt_required()
 def update_property(property_id):
     property = Property.query.filter_by(id=property_id).first()
 
-    if property:
+    user_id = jwt_get_identity()                     #current user id
+
+
+    if property.user_id == user_id:
         data = request.get_json()
 
         for attr in data:
@@ -202,14 +214,21 @@ def update_property(property_id):
 
 #delete property
 @prop_bp.route("/properties/<int:property_id>", methods=['DELETE'])
+@jwt_required()
 def delete_property(property_id):
     property = Property.query.filter_by(id=property_id).first()
 
     if property:
-        db.session.delete(property)
-        db.session.commit()
-        response = jsonify({"success": "Property deleted successfully"}), 200
-    else:
-        response = jsonify({"error": "Property not found"}), 404
+      if property.user_id == user_id:
 
-    return response
+         db.session.delete(property)
+         db.session.commit()
+         return jsonify({"success": "Property deleted successfully"}), 200
+
+      else:
+         return jsonify({"error":"unauthorised access"})
+
+    else:
+        return jsonify({"error": "Property not found"}), 404
+
+
